create 'models' package
